Update the syntax of the MIT license name
I'm finding that our automated license lister, https://github.com/Comcast/php-legal-licenses/, doesn't recognize the license on this project. I'm thinking it might be due to the way the name is stated. I'm proposing updating it to be the same as in a repo where the license is currently recognized.